Fix crash in Ficsit-Network interop code
This commit adds netFuncMeta_ methods for all Ficsit-Networks interop code in FGWiremodBuildable. These methods need to be called synchronously, so this commit adds netFuncMeta_ calls for all methods, and sets runtime to 0 (Synchronous) to prevent a crash due to Ficsit-Networks calling these methods from a background lua thread.
The crash happens when two FicsitWiremod components are connected together with a wire, and a setX method is called via FicsitNetworks code. setX functions attempt to redraw the wires, which leads to a crash due to wires being created on a background thread. This crash does not happen when the component that is called from a FicsitNetworks computer is NOT connected to another component with a FicsitWiremod wire.