Fix player hitbox blocking wiring tools raycast detection
Apply a 200-unit offset to raycast tests in the AWiremodBaseTool::GetTargetLookAt function

When moving around for long enough (potentially with a hypertube), the wiring tools will eventually have trouble detecting other entities besides the player entity. This fixes that by applying an (arbitrarily chosen) offset to the GetTargetLookAt raycast of 200 units. The raycast will skip the player bounding box, and correctly detect the intended look-at target